+++ elmyra.ip.access.epo/CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - ui [fix]: don't rebuild comments everytime list gets focus
 - ui, ux: improve post database wipe experience
+- ui: fix document color indicator when document is just added to basket, without any score or dismiss flag set
 
 0.20.0
 ======