+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,10 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- middleware, ui: "opaquelinks" subsystem on top of JSON Web Tokens
+
+0.20.1
+======
 - ui [fix]: don't rebuild comments everytime list gets focus
 - ui, ux: improve post database wipe experience
 - ui: fix document color indicator when document is just added to basket, without any score or dismiss flag set